+++ ricaun.Revit.DB/Revit/App.cs
@@ -1,0 +1,27 @@
+﻿using Autodesk.Revit.UI;
+using ricaun.Revit.UI;
+using System.Collections.Generic;
+
+namespace ricaun.Revit.DB.Revit
+{
+    [AppLoader]
+    public class App : IExternalApplication
+    {
+        private RibbonPanel ribbonPanel;
+        public Result OnStartup(UIControlledApplication application)
+        {
+            ribbonPanel = application.CreatePanel("DB");
+            ribbonPanel.CreatePushButton<Commands.Command>()
+                .SetLargeImage("/UIFrameworkRes;component/ribbon/images/revit.ico");
+            return Result.Succeeded;
+        }
+
+        public Result OnShutdown(UIControlledApplication application)
+        {
+            ribbonPanel?.Remove();
+            return Result.Succeeded;
+        }
+    }
+
+
+}